+++ src/FolkerKinzel.MimeTypes/MimeType_const.cs
@@ -14,7 +14,7 @@ public sealed partial class MimeType
     /// <summary>
     /// Minimum count of characters at which a line of an Internet Media Type <see cref="string"/> is wrapped.
     /// </summary>
-    public const int MinimumLineLength = 64;
+    public const int MinLineLength = 64;
 
 
     /// <summary>